opacity rules rewritten
I've changed the active opacity to be one, as it's annoying when I take
screenshots of something, and I can see my background through the
window.

I've also removed the opacity clauses to make things opaque before, as
now *all* active windows are opaque by default.

I've also added another clause, this time different, that reduces the
opacity to terminal windows. It uses the initialTitle as the actual
title changes with the program being run.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>